api/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-mail
mysql-connector-python
pyjwt
werkzeug